+++ src/rstr/stic_dis.cpp
@@ -84,7 +84,7 @@ extern uchar multy_language;
 
 extern int16_t pitchsize ;
 
-extern int16_t nIncline  ;
+extern int nIncline  ;
 extern uchar fax1x2;	// MK NEW 05.01.1993
 extern int  inc_num_EEM;	// in ST_TOOLS.C
 extern int  dis_LIMIT_EEM;	// in ST_TOOLS.C;
